Add configuration for minimum parallel pages in PDF mapping
This commit introduces a configuration setting for the minimum number of pages which triggers parallel processing in mapping of PDF pages to student names. This parameter, `pdf.mapper.min-parallel-pages`, has been added in `application.yml` and is used in `PdfPageStudentMapper` to decide the method of PDF processing, either parallel or non-parallel. This allows easier modification of the limit based on the capability of the system where the application is deployed.