Add passed/failed status to student report
Introduced a new function to the SubjectResponse class to check if the average mark is over 5.5 and the student has passed. This function is used in the StudentResponse class to gather all subjects that have not been passed. Result is reflected in the students report view in the form of an alert banner based on the calculated status. Relevant styles for the alert banners have been added to the styles.css.